+++ src/extension.ts
@@ -4,7 +4,6 @@ import * as vscode from 'vscode';
 import { CircularImportsProvider } from './circularImports';
 import { Core } from './core';
 import { OverviewProvider } from './overview';
-import { disposeStatusBarItem, initStatusBarItem } from './statusBarItem';
 import { TDependency } from './tdependency';
 import { showOutputWindow } from './unused-exports/log';
 import { UnusedExportsProvider } from './unusedExports';
@@ -16,8 +15,6 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.window.showInformationMessage('We cannot check an empty workspace!');
     return;
   }
-
-  initStatusBarItem();
 
   const cores = workspaceFolders.map(
     (wsf) => new Core(wsf.name, wsf.uri.fsPath)
@@ -149,6 +146,8 @@ export function activate(context: vscode.ExtensionContext) {
       Core.findInFile(filePath, unusedExportOrCircularImport)
   );
   context.subscriptions.push(disposable);
+
+  refreshAllCores(cores);
 }
 
 function refreshAllCores(cores: Core[]) {
@@ -156,6 +155,4 @@ function refreshAllCores(cores: Core[]) {
 }
 
 // find-unused-exports:ignore-next-line-exports
-export function deactivate() {
-  disposeStatusBarItem();
-}
+export function deactivate() {}