+++ src/refreshing.ts
@@ -1,0 +1,8 @@
+import { TDependency, DependencyType } from './tdependency';
+
+export const Refreshing = new TDependency(
+  undefined,
+  '-',
+  DependencyType.REFRESH,
+  'Refreshing...'
+);