+++ src/unused-exports/app.ts
@@ -16,10 +16,10 @@ import { getOnlyUsefullFiles } from './usefullFiles';
 const fixPath = (path: string, prefixLen: number): string =>
   fixPathSeparator(path.substring(prefixLen));
 
-const makePathRelativeToProject = (
+async function makePathRelativeToProject(
   relations: TRelation[],
   absPathToPrj: string
-): void => {
+): Promise<void> {
   const pathDelim = path.delimiter;
   const len = absPathToPrj.length + pathDelim.length;
   relations.forEach((r) => {
@@ -31,41 +31,41 @@ const makePathRelativeToProject = (
 
     r.imports.forEach((i) => (i.path = fixPath(i.path, len)));
   });
-};
+}
 
-export function app(
+export async function app(
   absPathToPrj: string,
   overviewContext: OverviewContext
-): TNotUsed[] {
+): Promise<TNotUsed[]> {
   const startTime = new Date();
 
   resetLog();
   log(startTime.toISOString());
   let ts = log('Path to project', pathResolve(absPathToPrj));
-  const context = makeContext(absPathToPrj, overviewContext);
-  const sourceFiles = getSourceFiles(absPathToPrj, context);
+  const context = await makeContext(absPathToPrj, overviewContext);
+  const sourceFiles = await getSourceFiles(absPathToPrj, context);
   ts = log('Finding the sources took', undefined, ts);
-  const parsedFiles = getParsedFiles(sourceFiles);
+  const parsedFiles = await getParsedFiles(sourceFiles);
   ts = log('Parsing the files took', undefined, ts);
-  const projectFiles = getOnlyProjectImports(context, parsedFiles);
+  const projectFiles = await getOnlyProjectImports(context, parsedFiles);
   ts = log('Processed files', projectFiles.length, ts);
-  const usefullFiles = getOnlyUsefullFiles(projectFiles);
+  const usefullFiles = await getOnlyUsefullFiles(projectFiles);
   ts = log('Files having imports|exports', usefullFiles.length, ts);
 
-  const imports = getImports(usefullFiles);
+  const imports = await getImports(usefullFiles);
   ts = log('Total imports', imports.length, ts);
-  const exports = getExports(usefullFiles, imports);
+  const exports = await getExports(usefullFiles, imports);
   ts = log('Total exports', exports.length, ts);
-  const relations = buildRelations(imports, exports, context.main);
-  makePathRelativeToProject(relations, absPathToPrj);
+  const relations = await buildRelations(imports, exports, context.main);
+  await makePathRelativeToProject(relations, absPathToPrj);
   ts = log('Analysed files', relations.length, ts);
-  const notUsed = getNotUsed(relations);
+  const notUsed = await getNotUsed(relations);
   const finalList = notUsed.sort(sortNotUsedFn);
   const numNotUsedExports = finalList.length;
   ts = log('Not used exports', numNotUsedExports, ts);
 
   const [unusedExportsAndCircularImportsList, numCircularImports] =
-    detectCircularImports(relations, finalList, ts);
+    await detectCircularImports(relations, finalList, ts);
 
   const endTime = new Date();
   const timeDiffMs: number = endTime.getTime() - startTime.getTime();