+++ src/unused-exports/fsUtils.ts
@@ -4,7 +4,10 @@ import * as path from 'path';
 import { OverviewContext } from '../overviewContext';
 import { log } from './log';
 
-export const readJsonFile = (path: string, overviewContext: OverviewContext): { [kes: string]: any } | undefined => {
+export const readJsonFile = (
+  path: string,
+  overviewContext: OverviewContext
+): { [kes: string]: any } | undefined => {
   if (fs.existsSync(path) === false) {
     return undefined;
   }
@@ -12,7 +15,10 @@ export const readJsonFile = (path: string, overviewContext: OverviewContext): { 
   try {
     let content = fs.readFileSync(path, 'utf8');
     /* we remove the comments from it */
-    content = content.replace(/\\"|"(?:\\"|[^"])*"|(\/\/.*|\/\*[\s\S]*?\*\/)/g, (m, g) => (g ? '' : m));
+    content = content.replace(
+      /\\"|"(?:\\"|[^"])*"|(\/\/.*|\/\*[\s\S]*?\*\/)/g,
+      (m, g) => (g ? '' : m)
+    );
     return JSON.parse(content);
   } catch (e: any) {
     log(`Error parsing "${path}"`, e.message ?? e);
@@ -63,7 +69,11 @@ export function fixDriverLetterCase(filePath: string): string {
   return driverLetter + restPath;
 }
 
-export function globSync(globRe: string, cwd: string = '.', globIgnore?: string[]): string[] {
+export function globSync(
+  globRe: string,
+  cwd: string = '.',
+  globIgnore?: string[]
+): string[] {
   const ignore = globIgnore?.map(path.posix.normalize);
   const res = glob.sync(globRe, {
     cwd,