+++ src/unused-exports/importedFiles.ts
@@ -1,22 +1,42 @@
 import * as path from 'path';
 import { TContext } from './context';
-import { fixDriverLetterCase, globSync, isDirectory, isFile, pathResolve } from './fsUtils';
+import {
+  fixDriverLetterCase,
+  globSync,
+  isDirectory,
+  isFile,
+  pathResolve,
+} from './fsUtils';
 import { log } from './log';
 import { TTsImport, TTsParsed } from './parsedFiles';
 
 const defaultModuleSuffixes = [''];
 
-export const getOnlyProjectImports = (context: TContext, parsedFiles: TTsParsed[]): TTsParsed[] => {
-  const { allowJs, baseUrl, moduleSuffixes = defaultModuleSuffixes, paths } = context;
+export async function getOnlyProjectImports(
+  context: TContext,
+  parsedFiles: TTsParsed[]
+): Promise<TTsParsed[]> {
+  const {
+    allowJs,
+    baseUrl,
+    moduleSuffixes = defaultModuleSuffixes,
+    paths,
+  } = context;
 
   parsedFiles.forEach((tsParsed) => {
     const { path: filePath, imports } = tsParsed;
-    const mapFn = makeImportAbs(baseUrl, path.dirname(filePath), moduleSuffixes, paths, allowJs);
+    const mapFn = makeImportAbs(
+      baseUrl,
+      path.dirname(filePath),
+      moduleSuffixes,
+      paths,
+      allowJs
+    );
     tsParsed.imports = imports.map(mapFn).filter(importValid);
   });
 
   return parsedFiles;
-};
+}
 
 function importValid(anImport: TTsImport | undefined): anImport is TTsImport {
   return anImport !== undefined;
@@ -91,7 +111,11 @@ const getGlobRegexp = (path: string, allowJs?: boolean): string =>
 const getDirGlobRegexp = (rootPath: string, allowJs?: boolean): string =>
   pathResolve(rootPath, allowJs ? `index.{ts,tsx,js,jsx}` : `index.{ts,tsx}`);
 
-function resolveFilePath(filePath: string, moduleSuffixes: string[], allowJs?: boolean): string | undefined {
+function resolveFilePath(
+  filePath: string,
+  moduleSuffixes: string[],
+  allowJs?: boolean
+): string | undefined {
   if (isFile(filePath)) {
     return filePath;
   }
@@ -99,7 +123,13 @@ function resolveFilePath(filePath: string, moduleSuffixes: string[], allowJs?: b
   try {
     /* try it as file */
     // const globReFile = getGlobRegexp(filePath, allowJs);
-    const resFile = doGlob('file', getGlobRegexp, filePath, moduleSuffixes, allowJs);
+    const resFile = doGlob(
+      'file',
+      getGlobRegexp,
+      filePath,
+      moduleSuffixes,
+      allowJs
+    );
     if (resFile) {
       return resFile;
     }
@@ -109,7 +139,13 @@ function resolveFilePath(filePath: string, moduleSuffixes: string[], allowJs?: b
     }
 
     /* try it as directory */
-    const resDir = doGlob('folder', getDirGlobRegexp, filePath, moduleSuffixes, allowJs);
+    const resDir = doGlob(
+      'folder',
+      getDirGlobRegexp,
+      filePath,
+      moduleSuffixes,
+      allowJs
+    );
     if (resDir) {
       return resDir;
     }
@@ -157,7 +193,10 @@ function doOneGlob(
     const res = globSync(globRe)?.[0];
     return res ? fixDriverLetterCase(res) : undefined;
   } catch (err: any) {
-    log(`Exception glob: cannot resolve path to '${filePath}'. Tried ${tryMode}`, err?.message || err);
+    log(
+      `Exception glob: cannot resolve path to '${filePath}'. Tried ${tryMode}`,
+      err?.message || err
+    );
     throw err;
   }
 }