+++ src/unused-exports/sourceFiles.ts
@@ -10,7 +10,10 @@ export interface TTsFile {
 
 const defaultExclude = ['**/node_modules/**/*', '**/*.d.ts'];
 
-export function getSourceFiles(pathToPrj: string, context: TContext): TTsFile[] {
+export async function getSourceFiles(
+  pathToPrj: string,
+  context: TContext
+): Promise<TTsFile[]> {
   const { allowJs, exclude = [], include, files } = context;
 
   /* A glob must use only forward slashes */
@@ -27,20 +30,36 @@ export function getSourceFiles(pathToPrj: string, context: TContext): TTsFile[] 
     context.overviewContext.globInclude = [globRegexp];
     context.overviewContext.globExclude = globExcludeExtended;
     context.overviewContext.numDefaultExclude = defaultExclude.length;
-    globFile(res, pathToPrj, globRegexp, globExcludeExtended, context.overviewContext);
+    globFile(
+      res,
+      pathToPrj,
+      globRegexp,
+      globExcludeExtended,
+      context.overviewContext
+    );
     return res;
   }
 
   /* We want to see the stats before doing the actions */
-  const includes = include ? globInclude.map((gi) => applyGlob(gi, globRegexp)) : [];
-  const includeGlobs: string[] = files ? [...explicitFiles, ...includes] : includes;
+  const includes = include
+    ? globInclude.map((gi) => applyGlob(gi, globRegexp))
+    : [];
+  const includeGlobs: string[] = files
+    ? [...explicitFiles, ...includes]
+    : includes;
   context.overviewContext.globInclude = includeGlobs;
   context.overviewContext.globExclude = globExclude;
   context.overviewContext.numDefaultExclude = undefined;
 
   const res: TTsFile[] = [];
   if (files !== undefined) {
-    globFiles(res, pathToPrj, explicitFiles, undefined, context.overviewContext);
+    globFiles(
+      res,
+      pathToPrj,
+      explicitFiles,
+      undefined,
+      context.overviewContext
+    );
   }
   if (include !== undefined) {
     globFiles(res, pathToPrj, includes, globExclude, context.overviewContext);
@@ -51,7 +70,8 @@ export function getSourceFiles(pathToPrj: string, context: TContext): TTsFile[] 
   return res;
 }
 
-const getGlobRegexp = (allowJs?: boolean): string => (allowJs ? '**/*.{ts,js}?(x)' : '**/*.ts?(x)');
+const getGlobRegexp = (allowJs?: boolean): string =>
+  allowJs ? '**/*.{ts,js}?(x)' : '**/*.ts?(x)';
 
 function fixPaths(paths: string[]): string[] {
   return paths.map((f) => fixPath(f));
@@ -122,5 +142,9 @@ function globFile(
     res.push({ path: source });
     count++;
   });
-  addGlobInclude(ctx, getAdjustedPath(fixPath(pathToFolder), globRegexp), count);
+  addGlobInclude(
+    ctx,
+    getAdjustedPath(fixPath(pathToFolder), globRegexp),
+    count
+  );
 }