+++ src/unused-exports/usefullFiles.ts
@@ -1,6 +1,11 @@
 import { TTsParsed } from './parsedFiles';
 
-export const getOnlyUsefullFiles = (files: TTsParsed[]): TTsParsed[] =>
-  files.filter((file) => isNotEmpty(file.imports) || isNotEmpty(file.exports));
+export async function getOnlyUsefullFiles(
+  files: TTsParsed[]
+): Promise<TTsParsed[]> {
+  return files.filter(
+    (file) => isNotEmpty(file.imports) || isNotEmpty(file.exports)
+  );
+}
 
 const isNotEmpty = (arr?: unknown[]): boolean => (arr ? arr.length > 0 : false);